requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
instructor
requests>=2.31.0